+++ lib/dtos/assessment_taker/assessment_taker_create_dto.dart
@@ -1,0 +1,13 @@
+import 'package:school_erp/dtos/dto.dart';
+
+class AssessmentTakerCreateDTO extends CreateDTO {
+  String? assessmentId;
+  final String subjectYearId;
+  final String sectionId;
+
+  AssessmentTakerCreateDTO({
+    required this.subjectYearId,
+    required this.sectionId,
+    this.assessmentId,
+  });
+}
